+++ models/hatebert_baseline.py
@@ -33,7 +33,7 @@ class HateBERTBaseline(nn.Module):
         print(f"Loading HateBERT: {model_name}")
         base_model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            dtype=torch.float16,
+            torch_dtype=torch.bfloat16,
             attn_implementation="eager",
         )
 
@@ -50,7 +50,7 @@ class HateBERTBaseline(nn.Module):
 
         # Classification head on top of last token hidden state
         hidden_size = self.model.config.hidden_size
-        self.classifier = nn.Linear(hidden_size, num_labels)
+        self.classifier = nn.Linear(hidden_size, num_labels).float()
         self.num_labels  = num_labels
 
         trainable = sum(p.numel() for p in self.parameters() if p.requires_grad)
